Create separate controllers from table and database structure pages
